move IsValidSlackChannel to services/webhook package
